Added .gitattributes file to force LF checkout on DOS systems
Most windows machines are configured to normalize line endings to CRLF at checkout. 
This is deadly for people trying to debug the dockerfile on docker for windows, because docker deploys files as is.
What happens is that at build time this does not cause any issues yet. but when trying to actually run the image it becomes clear that absolutely nothing works and the image exits right at the start.